+++ src/lib_injector/injector_server.ml
@@ -102,6 +102,8 @@ module Parameters :
     | Never_included ->
         (* Forget operations that are never included *)
         return Forget
+
+  let metrics_registry = Prometheus.CollectorRegistry.create ()
 end
 
 include Injector_functor.Make (Parameters)