+++ src/lib_smart_rollup_node/node_context.ml
@@ -285,6 +285,7 @@ let set_l2_head node_ctxt (head : Sc_rollup_block.t) =
   let open Lwt_result_syntax in
   let+ () = Store.L2_head.write node_ctxt.store.l2_head head in
   notify_processed_tezos_level node_ctxt head.header.level ;
+  Metrics.wrap (fun () -> Metrics.Inbox.set_head_level head.header.level) ;
   Lwt_watcher.notify node_ctxt.global_block_watcher head
 
 let is_processed {store; _} head = Store.L2_blocks.mem store.l2_blocks head
@@ -583,7 +584,10 @@ let set_lcc node_ctxt lcc =
   let open Lwt_result_syntax in
   let lcc_l1 = Reference.get node_ctxt.lcc in
   let* () = Store.Lcc.write node_ctxt.store.lcc lcc in
-  if lcc.level > lcc_l1.level then Reference.set node_ctxt.lcc lcc ;
+  Metrics.Info.set_lcc_level_local lcc.level ;
+  if lcc.level > lcc_l1.level then (
+    Reference.set node_ctxt.lcc lcc ;
+    Metrics.Info.set_lcc_level_l1 lcc.level) ;
   let*! () =
     Commitment_event.last_cemented_commitment_updated lcc.commitment lcc.level
   in
@@ -615,6 +619,8 @@ let register_published_commitment node_ctxt commitment ~first_published_at_level
         {first_published_at_level; published_at_level}
     else return_unit
   in
+  if published_by_us then Metrics.Info.set_lpc_level_local level
+  else Metrics.Info.set_lpc_level_l1 level ;
   when_ published_by_us @@ fun () ->
   let* () = Store.Lpc.write node_ctxt.store.lpc commitment in
   let update_lpc_ref =
@@ -1021,11 +1027,14 @@ let gc ?(wait_finished = false) ?(force = false) node_ctxt ~(level : int32) =
             Block_hash.pp
             hash
       | Some {context; _} ->
+          let start_timestamp = Time.System.now () in
           let* gc_lockfile =
             Utils.lock (gc_lockfile_path ~data_dir:node_ctxt.data_dir)
           in
           let*! () = Event.calling_gc ~gc_level ~head_level:level in
           let*! () = save_gc_info node_ctxt ~at_level:level ~gc_level in
+          Metrics.wrap (fun () ->
+              Metrics.GC.set_oldest_available_level gc_level) ;
           (* Start both node and context gc asynchronously *)
           let*! () = Context.gc node_ctxt.context context in
           let* () = Store.gc node_ctxt.store ~level:gc_level in
@@ -1033,6 +1042,10 @@ let gc ?(wait_finished = false) ?(force = false) node_ctxt ~(level : int32) =
             let open Lwt_syntax in
             let* () = Context.wait_gc_completion node_ctxt.context
             and* () = Store.wait_gc_completion node_ctxt.store in
+            Metrics.wrap (fun () ->
+                let stop_timestamp = Time.System.now () in
+                Metrics.GC.set_process_time
+                @@ Ptime.diff stop_timestamp start_timestamp) ;
             let* () = Event.gc_finished ~gc_level ~head_level:level in
             Utils.unlock gc_lockfile
           in